+++ resolwe_bio/variants/serializers.py
@@ -6,8 +6,6 @@ Variants Serializers
 
 """
 
-from django.db import transaction
-
 from resolwe.flow.serializers import ResolweBaseSerializer
 
 from resolwe_bio.variants.models import (
@@ -28,7 +26,7 @@ class VariantTranscriptSerializer(ResolweBaseSerializer):
         model = VariantAnnotationTranscript
         fields = [
             "id",
-            "variant_annotation_id",
+            "variant_annotation",
             "annotation",
             "annotation_impact",
             "gene",
@@ -48,7 +46,7 @@ class VariantAnnotationSerializer(ResolweBaseSerializer):
 
         model = VariantAnnotation
         read_only_fields = ("id",)
-        update_protected_fields = ("variant_id",)
+        update_protected_fields = ("variant",)
         fields = (
             read_only_fields
             + update_protected_fields
@@ -61,36 +59,6 @@ class VariantAnnotationSerializer(ResolweBaseSerializer):
                 "transcripts",
             )
         )
-
-    @transaction.atomic
-    def create(self, validated_data):
-        """Create a new VariantAnnotation instance."""
-        transcripts_data = validated_data.pop("transcripts", None)
-        annotation = VariantAnnotation.objects.create(**validated_data)
-        if transcripts_data:
-            VariantAnnotationTranscript.objects.bulk_create(
-                VariantAnnotationTranscript(
-                    **transcript_data, variant_annotation=annotation
-                )
-                for transcript_data in transcripts_data
-            )
-        return annotation
-
-    @transaction.atomic
-    def update(self, instance, validated_data):
-        """Update the variant annotation."""
-        transcripts_data = validated_data.pop("transcripts", None)
-        annotation = super().update(instance, validated_data)
-        if transcripts_data is not None:
-            # Always create new transcripts since updating logic is tedious.
-            annotation.transcripts.all().delete()
-            VariantAnnotationTranscript.objects.bulk_create(
-                VariantAnnotationTranscript(
-                    **transcript_data, variant_annotation=annotation
-                )
-                for transcript_data in transcripts_data
-            )
-        return annotation
 
 
 class VariantSerializer(ResolweBaseSerializer):
@@ -112,41 +80,6 @@ class VariantSerializer(ResolweBaseSerializer):
             "alternative",
             "annotation",
         )
-
-    @transaction.atomic
-    def create(self, validated_data):
-        """Create a new Variant instance."""
-        annotation_data = validated_data.pop("annotation", None)
-        variant = Variant.objects.create(**validated_data)
-        if annotation_data:
-            VariantAnnotationSerializer().create(
-                {"variant_id": variant.id, **annotation_data}
-            )
-        return variant
-
-    @transaction.atomic
-    def update(self, instance, validated_data):
-        """Update the Variant instance."""
-        delete_annotation = (
-            "annotation" in validated_data and validated_data["annotation"] is None
-        )
-        annotation_data = validated_data.pop("annotation", None)
-        variant = super().update(instance, validated_data)
-        if annotation_data:
-            # Update existing annotation.
-            if instance.has_annotation:
-                VariantAnnotationSerializer().update(
-                    instance.annotation, annotation_data
-                )
-            # Create new annotation.
-            else:
-                VariantAnnotationSerializer().create(
-                    {"variant_id": variant.id, **annotation_data}
-                )
-        elif delete_annotation and instance.annotation:
-            instance.annotation.delete()
-            instance.annotation = None
-        return variant
 
 
 class VariantCallSerializer(ResolweBaseSerializer):
@@ -191,11 +124,9 @@ class VariantExperimentSerializer(ResolweBaseSerializer):
 
     def perform_create(self, serializer):
         """Set the contributor to the current user."""
-        print("perform create", serializer.validated_data)
         serializer.save(contributor=self.context["request"].user)
         super().perform_create(serializer)
 
     def create(self, validated_data):
         """Create a new VariantExperiment instance."""
-        print("Creating, validated_data", validated_data)
         return super().create(validated_data)